Made front end with react and react-bootstrap
Created three unique components: NavMenu.tsx, MainContent.tsx, Task.tsx. Nav menu is used to display the navbar on top of the site. Main content is for the site's main content (main page). Task is for the individual tasks (present in section 3 in MainContent.tsx). Also, there are two folder, on for assets (images) and one for the custom font used.