Created an authentication system with tokens so whenever a user logs in, the system can save this information
Whenever a user logs in, a token is generated and given from the Backend. Each component can now access this information (that a token has been created) and handle it however it needs. At last, changed the routing of the components and used react-router's navigate hook to redirect users whenever required. More specifically, after a user loggs in they are redirected to /profile, and if a user has not logged in, /profile redirects them to /login